Finish Adding books and listing books and labels
