+++ Classes/list_books.rb
@@ -1,0 +1,10 @@
+require_relative './list_data'
+
+class ListBooks < ListData
+  def list_books(data)
+    return puts "No Books found\n\n" if data.empty?
+
+    data.select! { |item| item.instance_of?(Book) }
+    data.each_with_index { |book, index| get_properties(book, index, ['Publisher'], [book.publisher]) }
+  end
+end
